Move nav buttons into there own component in the nav bar
