+++ src/app/components/map/nav-bar/nav-bar.component.html
@@ -9,87 +9,11 @@
   </div>
 
   <div class="layout">
-
-  <div *ngIf="products" matTooltip="Download Queue">
-    <button
-      style="background: white"
-      class="spacing"
-      (click)="onOpenDownloadQueue()"
-      [disabled]="products.length <= 0" color="basic"
-      mat-fab>
-        <mat-icon
-          [matBadge]="products.length"
-          matBadgePosition="below"
-          matBadgeColor="{{ products.length > 0 ? 'accent' : 'primary' }}"
-          matBadgeOverlap="false">
-            cloud_download
-        </mat-icon>
-    </button>
-  </div>
-
-  <button
-    (click)="onAccountButtonClicked()"
-    matTooltip="Login"
-    class="spacing"
-    color="basic"
-    mat-fab>
-    <mat-icon>{{ datapoolAuthService.isLoggedIn ? 'check_circle' : 'account_circle' }}</mat-icon>
-  </button>
-
-  <button
-    matTooltip="Share Search"
-    class="spacing"
-    color="basic"
-    [matMenuTriggerFor]="shareMenu"
-    mat-fab>
-      <mat-icon>share</mat-icon>
-  </button>
-
-  <mat-menu #shareMenu="matMenu">
-    <button mat-menu-item (click)="onCopy()"
-      matSuffix>
-      <mat-icon>file_copy</mat-icon>
-      <span>Copy Search Link</span>
-    </button>
-    <a (click)="onShareWithEmail()"
-      href="mailto:?subject=&body=:%20" target="_blank"
-      title="Send email"
-      mat-menu-item matSuffix>
-        <mat-icon>email</mat-icon>
-        <span>Share With Email</span>
-    </a>
-  </mat-menu>
-
-  <button
-    matTooltip="ASF Apps"
-    class="spacing"
-    color="basic"
-    [matMenuTriggerFor]="appMenu"
-    mat-fab>
-      <mat-icon>apps</mat-icon>
-  </button>
-  <mat-menu #appMenu="matMenu">
-    <a mat-menu-item [href]="asfWebsiteUrl" target="_blank">
-      ASF Website
-    </a>
-    <a mat-menu-item href="https://baseline.asf.alaska.edu" target="_blank">
-      Baseline Tool
-    </a>
-    <a mat-menu-item href="{{ asfWebsiteUrl }}/get-data/api" target="_blank">
-     ASF API
-    </a>
-    <button mat-menu-item [matMenuTriggerFor]="thematicDatasets">Thematic Datasets</button>
-  </mat-menu>
-
-  <mat-menu #thematicDatasets="matMenu">
-    <a mat-menu-item href="{{ asfWebsiteUrl }}/other-data/gismo" target="_blank">GISMO</a>
-    <a mat-menu-item href="{{ asfWebsiteUrl }}/other-data/glacier-speed" target="_blank">Glacier Speed</a>
-    <a mat-menu-item href="{{ asfWebsiteUrl }}/other-data/polar-year-07-08" target="_blank">Polar Year</a>
-    <a mat-menu-item href="{{ asfWebsiteUrl }}/other-data/ramp" target="_blank">RAMP</a>
-    <a mat-menu-item href="{{ asfWebsiteUrl }}/other-data/sea-ice" target="_blank">Sea Ice MEaSUREs</a>
-    <a mat-menu-item href="{{ asfWebsiteUrl }}/wetlands" target="_blank">Wetlands MEaSUREs</a>
-  </mat-menu>
-
+    <app-nav-buttons
+      (openQueue)="onOpenDownloadQueue()"
+      [isSideMenuOpen]="isSideMenuOpen"
+      [products]="products"
+      ></app-nav-buttons>
   </div>
 </div>
