+++ src/app/components/map/nav-bar/nav-bar.component.scss
@@ -13,7 +13,3 @@
   display: flex;
   justify-content: flex-end;
 }
-
-.spacing {
-  margin: 4px;
-}
