+++ src/app/components/map/nav-bar/nav-bar.component.ts
@@ -5,12 +5,7 @@ import {
   animate, transition
 } from '@angular/animations';
 
-import { ClipboardService } from 'ngx-clipboard';
-
 import * as models from '@models';
-
-import { DatapoolAuthService } from '@services';
-import { environment } from '@environments/environment';
 
 @Component({
   selector: 'app-nav-bar',
@@ -29,7 +24,6 @@ import { environment } from '@environments/environment';
   ],
 })
 export class NavBarComponent {
-  public asfWebsiteUrl = 'https://www.asf.alaska.edu';
 
   @Output() openQueue = new EventEmitter<void>();
   @Output() doSearch = new EventEmitter<void>();
@@ -37,39 +31,11 @@ export class NavBarComponent {
   @Input() products: models.Sentinel1Product[];
   @Input() isSideMenuOpen: boolean;
 
-  constructor(
-    public datapoolAuthService: DatapoolAuthService,
-    public clipboard: ClipboardService,
-  ) {}
-
   public onOpenDownloadQueue(): void {
     this.openQueue.emit();
   }
 
   public onDoSearch(): void {
     this.doSearch.emit();
-  }
-
-  public onAccountButtonClicked() {
-    if (!this.datapoolAuthService.isLoggedIn) {
-      this.datapoolAuthService.login();
-    } else {
-      this.datapoolAuthService.profileInfo()
-        .subscribe(console.log);
-      console.log('user is logged in...');
-    }
-  }
-
-  public onCopy(): void {
-    this.clipboard.copyFromContent(window.location.href);
-  }
-
-  public onShareWithEmail() {
-    const subject = `New Search - ${encodeURIComponent(document.title)}`;
-
-    window.open(
-      `mailto:?subject=${subject}` +
-      `&body=${encodeURIComponent(document.URL)}`
-    );
   }
 }
