+++ src/app/components/map/nav-bar/nav-buttons/nav-buttons.component.ts
@@ -1,4 +1,9 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
+
+import { ClipboardService } from 'ngx-clipboard';
+
+import { DatapoolAuthService } from '@services';
+import { Sentinel1Product } from '@models';
 
 @Component({
   selector: 'app-nav-buttons',
@@ -6,10 +11,45 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./nav-buttons.component.css']
 })
 export class NavButtonsComponent implements OnInit {
+  public asfWebsiteUrl = 'https://www.asf.alaska.edu';
 
-  constructor() { }
+  @Input() isSideMenuOpen: boolean;
+  @Input() products: Sentinel1Product[];
+
+  @Output() openQueue = new EventEmitter<void>();
+
+  constructor(
+    public datapoolAuthService: DatapoolAuthService,
+    public clipboard: ClipboardService,
+  ) {}
 
   ngOnInit() {
   }
 
+  public onOpenDownloadQueue(): void {
+    this.openQueue.emit();
+  }
+
+  public onAccountButtonClicked() {
+    if (!this.datapoolAuthService.isLoggedIn) {
+      this.datapoolAuthService.login();
+    } else {
+      this.datapoolAuthService.profileInfo()
+        .subscribe(console.log);
+      console.log('user is logged in...');
+    }
+  }
+
+  public onCopy(): void {
+    this.clipboard.copyFromContent(window.location.href);
+  }
+
+  public onShareWithEmail() {
+    const subject = `New Search - ${encodeURIComponent(document.title)}`;
+
+    window.open(
+      `mailto:?subject=${subject}` +
+      `&body=${encodeURIComponent(document.URL)}`
+    );
+  }
 }
