+++ src/app/components/map/nav-bar/nav-buttons/nav-buttons.module.ts
@@ -1,11 +1,20 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+
+import { MatBadgeModule, MatMenuModule, MatInputModule } from '@angular/material';
+import { MatSharedModule } from '@shared';
+
 import { NavButtonsComponent } from './nav-buttons.component';
 
 @NgModule({
   declarations: [NavButtonsComponent],
   imports: [
-    CommonModule
+    CommonModule,
+
+    MatBadgeModule,
+    MatMenuModule,
+    MatInputModule,
+    MatSharedModule,
   ],
   exports: [NavButtonsComponent],
 })
