+++ src/app/components/sidebar/search/search-selector/detailed-search-selector/detailed-search-selector.component.html
@@ -61,30 +61,4 @@
   <mat-card class="mat-elevation-z0">
     <img class="img-nasa-logo" src="/assets/meatball.png">
   </mat-card>
-<!--  <div style="margin-top: 7px; padding: 0 16px">-->
-<!--  <mat-accordion>-->
-<!--    <mat-expansion-panel>-->
-<!--      <mat-expansion-panel-header>-->
-<!--        <mat-panel-title>-->
-<!--          What are these search types?-->
-<!--        </mat-panel-title>-->
-<!--        <mat-panel-description>-->
-<!--        </mat-panel-description>-->
-<!--      </mat-expansion-panel-header>-->
-<!--        <h3>Data Set Search (most common)</h3>-->
-<!--        <p>-->
-<!--          Search for products from a dataset using a area of interest and filter based-->
-<!--          on date, dataset, product type and other metadata parameters.-->
-<!--        </p>-->
-<!--        <h3>List Search</h3>-->
-<!--        <p>-->
-<!--          Search using a list of granule names or a list of product names.-->
-<!--        </p>-->
-<!--        <h3>Mission Search</h3>-->
-<!--        <p>-->
-<!--          Search based on a mission, need a better description for this!-->
-<!--        </p>-->
-<!--    </mat-expansion-panel>-->
-<!--  </mat-accordion>-->
-<!--  </div>-->
 </mat-card>
